+++ src/base/os_exec_op.cc
@@ -17,7 +17,7 @@ namespace chef {
     const int MAX_LEN = 1024;
     char line[MAX_LEN] = {0};
     while(fgets(line, MAX_LEN, fp) != NULL) {
-      int rend = strlen(line) - 1;
+      size_t rend = strlen(line) - 1;
       if (line[rend] == '\n') {
         line[rend] = '\0';
       }
