+++ src/base/this_proc_op.h
@@ -24,37 +24,37 @@ namespace chef {
        * 进程启动时的unix时间戳，单位秒
        *
        */
-      static int boot_timestamp();
+      static int64_t boot_timestamp();
 
       /**
        * 进程至今运行时间，单位秒
        *
        */
-      static int up_duration_seconds();
+      static int64_t up_duration_seconds();
 
       /**
        * @return 成功返回进程当前线程数，失败返回0
        *
        */
-      static int num_of_threads();
+      static int32_t num_of_threads();
 
       /**
        * 内存管理的最小单位：物理页面的大小，单位字节
        *
        */
-      static int page_size();
+      static int64_t page_size();
 
       /**
        * 进程当前虚拟内存大小，对应top中的VIRT列，单位KBytes
        *
        */
-      static int virt_kb();
+      static int32_t virt_kb();
 
       /**
        * 进程当前常驻内存大小，对应top中的RES列，单位KBytes
        *
        */
-      static int res_kb();
+      static int32_t res_kb();
 
       /// @return e.g. /home/chef/starry-night/build/base_test/stuff_test
       static std::string exe_filepath();
@@ -71,13 +71,13 @@ namespace chef {
       /// content of /proc/self/stat
       static std::string stat();
 
-      static int pid();
+      static int32_t pid();
 
-      static int ppid();
+      static int32_t ppid();
 
-      static int uid();
+      static int32_t uid();
 
-      static int euid();
+      static int32_t euid();
 
       /// user name of os
       static std::string user_name();
