+++ src/base/thread_pool.h
@@ -42,7 +42,7 @@ namespace chef {
       void add(const task &t);
 
       ///
-      int num_of_undone_task();
+      uint64_t num_of_undone_task();
 
     private:
       void run_in_thread(int index);
